may the codex spin

+++ src/main/java/eladkay/quaeritum/api/spell/render/CircleSymbolInstruction.java
@@ -23,7 +23,7 @@ public class CircleSymbolInstruction implements ISymbolInstruction {
 
     @Override
     @SideOnly(Side.CLIENT)
-    public void render(EnumSpellElement element, int x, int y) {
+    public void render(EnumSpellElement element, float x, float y) {
         Tessellator tess = Tessellator.getInstance();
         VertexBuffer buffer = tess.getBuffer();
         int color = element.color();
