+++ src/main/java/eladkay/quaeritum/api/spell/render/NGonSymbolInstruction.java
@@ -34,7 +34,7 @@ public class NGonSymbolInstruction implements ISymbolInstruction {
 
     @Override
     @SideOnly(Side.CLIENT)
-    public void render(EnumSpellElement element, int x, int y) {
+    public void render(EnumSpellElement element, float x, float y) {
         Tessellator tess = Tessellator.getInstance();
         VertexBuffer buffer = tess.getBuffer();
         int color = element.color();
